+++ packages/flutter/lib/src/painting/text_painter.dart
@@ -579,7 +579,7 @@ class TextPainter {
           newWidth = maxIntrinsicWidth;
           break;
       }
-      newWidth = newWidth.clamp(minWidth, maxWidth) as double; // ignore: unnecessary_cast
+      newWidth = newWidth.clamp(minWidth, maxWidth);
       if (newWidth != _applyFloatingPointHack(_paragraph!.width)) {
         _paragraph!.layout(ui.ParagraphConstraints(width: newWidth));
       }
